requirements.txt
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
pyttsx3==2.90
python-dotenv==1.0.0
# Optional dependencies for macOS
# pyobjc-framework-Cocoa==10.1  # Only needed on macOS
# pyobjc-framework-AVFoundation==10.1  # Only needed on macOS 